UPSTREAM: amd/stoneyridge: Refactor SMI handler
Rewrite the handler to be more compact and extendable.  The old
functionality is duplicated after the rewrite.  All SMI source registers
(except for SmiSciStatus) behave identically so these are consolidated.
Register 0x80 contains sources 0-31, 0x81 sources 32-63, and so on.

Create a table of mini-handlers to be supported in the soc directory.
As SMI sources are discovered, attempt to find the corresponding handler
and then execute it.

BUG=none
BRANCH=none
TEST=none

Change-Id: I30d704a90f225a504714be770275ef3c12de68ad
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 36a2356fb1bb85468de1d9e191df890c46ff3861
Original-Change-Id: Ic7050ecf65c2af036fe297f429a0bbdc709ad4c1
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21746
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Marc Jones <marc@marcjonesconsulting.com>
Reviewed-on: https://chromium-review.googlesource.com/699391
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>