UPSTREAM: nb/intel/pineview: Port ACPI opregion to pineview
Port the ACPI opregion implementation that resides in
drivers/intel/gma to older platforms. It allows to include a vbt.bin and
allows GNU/Linux to load the opregion as ASLB is being set.

Windows' Intel will likely ignore it as it relies on legacy VBIOS
to be loaded at 0xc0000.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia8b7edffee3170691e29d6720b2ff1fb38b8cc6e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bb98b38b93d1f01fd3cd71c9e27ddbf4d3f1d244
Original-Change-Id: Ifc9fc52d84dcbb0da577e61467ece8a48752f44b
Original-Signed-off-by: Patrick Rudolph <siro@das-labor.org>
Original-Reviewed-on: https://review.coreboot.org/21775
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/733079
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>