UPSTREAM: drivers/uart8250mem: Check for zero base address
Before adding a new UART to the coreboot/lb table, verify that it
has a non-zero base address.  This is consistent with all other
functions that use the uart_platform_base() function.

This was tested on google/kahlee by using an invalid UART number
and forcing the base address to 0.  Execution was able to complete
through depthcharge and into the OS.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iba1a674d327ea8b1efc1745b4bd3e70c3e37317a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 78130663e5cac7939ee04f97ac1c7fcecde23820
Original-Change-Id: I6d8183a461f0fedc254bf88de5ec96629a2a80ef
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21996
Original-Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/720945
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>