UPSTREAM: intel/i210: Set bus master bit in command register
There are still Ethernet drivers out there in the wild which expect
that the master enable bit was set by firmware. A missing master
enable bit will lead to a non-functional driver. Though it is clear the
task of the driver to set this bit it is too late now. So work around
this issue on firmware level...again!

BUG=none
BRANCH=none
TEST=none

Change-Id: I8c7182e63111e093d88242a496758e7932bf5ef0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bd31642ad8cc8302e2455a072a745dad9ea579b9
Original-Change-Id: I677b22c643b73634b1a2129d948b991446e1f8fd
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/22036
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/733058
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>