UPSTREAM: src/mainboard/intel/glkrvp: Fix Lid switch support
SCI trigger logic had to be inverted.
This patch enables the system to wake up from S3 when lid is opened
when the system is in suspend state.We are trying to match what a
external EC card running ChromeEC FW is sending on the signal.

TEST = Verified that system wakes up from S3 on toggling lid switch
back to open state.

BUG=none
BRANCH=none
TEST=none

Change-Id: I414ad6fa2cda5d17ebc5e732b20fef07d9fdc197
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d038d53a33e4dae29db82b2dcb276ff5ae9d4a14
Original-Change-Id: Ib42a38088ee028eddc6769921b0552c569da25a9
Original-Signed-off-by: Shaunak Saha <shaunak.saha@intel.com>
Original-Reviewed-on: https://review.coreboot.org/22117
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Maximilian Schander <coreboot@mimoja.de>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/758246
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>