UPSTREAM: drv/intel/gma/opregion: fix opregion version for Windows
Although Intel's current spec[1] shows the OpRegion structure version
as being the top 16 bits of the field, Intel's Windows drivers
require the OpRegion structure version to be in the top 8 bits of
the field when not using a VGA BIOS (eg, NGI or GOP driver).

As the Linux i915 driver only checks that the version is >= 2,
there is no change in functionality there.

This change effectively matches Intel's implementation in TianoCore,
where the version is set to 0x0200 << 16.

[1] https://01.org/sites/default/files/documentation/skl_opregion_rev0p5.pdf

TEST: Boot Windows [8.1,10] in UEFI mode w/GOP graphics init and
Legacy mode with libgfxinit, observe Intel GPU driver functional.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ifbf769408313c7d62d4a88a7eb6607de4737f302
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fd0a891ee9ed0947123e893505bdfc8ada408ab9
Original-Change-Id: Ic2903ee4829689ec4117aec93dce0b87cec6f313
Original-Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/22118
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/734046
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>