UPSTREAM: soc: Add Kconfig for each soc vendor
Allows explicit ordering for vendors that share a common configuration
that must be sourced last.

The issue is that chips in soc/{amd,intel}/[ab].* will be able to
override defaults set in this file, but Kconfig files that get sourced
later (soc/amd/[d-z].*) will NOT be able to override these defaults.

Note: intel and amd soc chips now need to be added manually to the new
Kconfig file

BUG=b:62235314
TEST=make lint-stable

Change-Id: Ibc627969b42ac2ffd0c9e18c3c1848da76ebf602
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: aa8e5d36b1b8a5ce1574c3562416e8488ebbb8cc
Original-Change-Id: Ida82ef184712e092aec1381a47aa1b54b74ed6b6
Original-Signed-off-by: Chris Ching <chingcodes@google.com>
Original-Reviewed-on: https://review.coreboot.org/22123
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/734048
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>