UPSTREAM: nb/intel/sandybridge/raminit: Fix SMBIOS 17 bus width
The bus width has to be encoded where the lower 3 bits are the bus width
in multiple of 8 and the following two bits give the error checking
bits in multiple of 8.
Hardcode to 64 bit as done on haswell.
TODO: Make it dynamic once there's ECC support.

BUG=none
BRANCH=none
TEST=none

Change-Id: I480388ac4219e745f32f7659b8555729805bd763
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5af2deae92abd1258a752f29fbe48c91458706c2
Original-Change-Id: I3b83a098205455b1c820d0436c6984938f261466
Original-Signed-off-by: Patrick Rudolph <siro@das-labor.org>
Original-Reviewed-on: https://review.coreboot.org/22261
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1155511
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>