UPSTREAM: rockchip/rk3399: mipi: properly configure PHY timing
These values are specified as constant time periods but the PHY
configuration is in terms of the current lane byte clock so using
constant values guarantees that the timings will be outside the
specification with some display configurations.

Derive the necessary configuration from the byte clock in order to
ensure that the PHY configuration is correct.

BUG=none
BRANCH=none
TEST=none

Change-Id: I293937e63e8d6d1e4236ec646be426f63f65ad02
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 45f1b01324ed1712092e80fed7e03fe088452729
Original-Change-Id: I396029956730907a33babe39c6a171f2fcea9dcd
Original-Signed-off-by: Lin Huang <hl@rock-chips.com>
Original-Reviewed-on: https://review.coreboot.org/22470
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Brian Norris <briannorris@chromium.org>
Original-Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/796224
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>