UPSTREAM: sb/intel/i82371eb: Rework ACPI tables
Rework ACPI tables based on a mix of previous work on asus/p2b,
other boards in tree with better ACPI support, and OEM BIOS.
To be pulled in by DSDTs of mainboards using this southbridge.

Disable on-the-fly generation of mainboard _CRS node. It is not working
as it should and causes runtime errors when booting Linux. This node
to be included in mainboard DSDTs in followup patches.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id0f4c0aec9f641f1b89b35f33e2d08dedce3528b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 97041de909b5468c1a62afd5e38bf59c6b5c7881
Original-Change-Id: Idda424de7859a36e4cac168d5469f9365a6ad421
Original-Signed-off-by: Keith Hui <buurin@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21671
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/816937
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>